Refactor Analytics and SessionManager for anonymous ID handling
* Introduce getCookie and generateRandomToken utility functions for cookie management and token generation
* Implement maybeSetAnonId method in Analytics to set a first-party cookie if not present
* Update SessionManager to utilize the new getCookie utility and remove redundant methods
* Enhance session initialization by ensuring anonymous ID is set correctly
